requirements.txt
# aws cdk
aws-cdk.core==1.182.0
aws-cdk.aws_lambda==1.182.0